requirements-test.txt
-e .

mock
tox
pytest
pytest-cov
pytest-django
pytest-flake8
djangorestframework
